requirements.txt
# Hair Style AI Generator - 商用レベル要件
# Core Flask Framework
Flask==3.0.3
Flask-SocketIO==5.3.6
Flask-WTF==1.2.1
Flask-Limiter==3.5.0

# 非同期処理・タスクキュー
Celery==5.3.6
Redis==5.0.1
eventlet==0.35.2
kombu==5.3.5

# AI API統合
google-genai>=1.0.0
requests==2.31.0

# 画像処理
Pillow==10.0.1

# ユーティリティ
python-dotenv==1.0.0
uuid

# 本番環境
gunicorn[eventlet]==21.2.0

# テスト・開発
pytest==7.4.3
pytest-flask==1.3.0
pytest-mock==3.12.0
pytest-cov==4.1.0
locust==2.17.0
coverage==7.3.0

# セキュリティ
cryptography>=41.0.0 